+++ CHANGELOG.md
@@ -24,6 +24,13 @@
   - `NEVec::new()`
   - `NEVec::first()`
 - **BREAKING:** non-empty maps and sets now behave similarly to their possibly empty counter parts: when created from an iterator with duplicates, the last occurence is kept.
+- **BREAKING:** Consistent API, new naming to align with Rust's naming conventions and indicate the fallibility of the function:
+  - `from_vec` to `try_from_vec`
+  - `from_map` to `try_from_map`
+  - `from_slice` to `try_from_slice`
+  - `from_set` to `try_from_set`.
+- **BREAKING:** `IteratorExt` is removed in favor of `IntoIteratorExt`. Now it's possible to call `try_into_nonempty_iter()` instead of `to_nonempty_iter()` on all regular iterators because regular iterators also implement `IntoIterator`.
+- **BREAKING:** `.iter()`, `.iter_mut()`, etc, are now prefixed with `nonempty_`
 - `FromNonEmptyIterator<T>` is now implemented for `HashSet<T, S>` instead of `HashSet<T>` (with the default hasher).
 
 #### Fixed
@@ -45,6 +52,7 @@
  - The rust version to which the library is build is now pinned, to avoid accidental breakage.
  - A [`justfile`](https://github.com/casey/just) that allows to run pre-configured commands to check the codebase. E.g. `just lint` or `just test`.
  - Benchmarks for `Vec` versus `NEVec`.
+ - Added `.iter()` methods to all collections returning a regular `Iterator`.
 
 ## 0.2.9 (2024-08-26)